Clean up info->decoded_size
Now non-predecoder-friendly formats set it to 0, and I also redefined
it as a size_t, because why the hell wasn't it before?

+++ clownaudio/decoders/common.h
@@ -20,7 +20,7 @@ typedef struct DecoderInfo
 {
 	unsigned int sample_rate;
 	unsigned int channel_count;
-	unsigned long decoded_size;
+	size_t decoded_size;
 	DecoderFormat format;
 } DecoderInfo;
